pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "slm-council"
version = "0.1.0"
description = "SLM Coding Council – Many-to-One autonomous coding engine on GCP"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Anmol" }]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.28.0",
    "google-cloud-aiplatform>=1.74.0",
    "google-auth>=2.36.0",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
    "rich>=13.9.0",
    "sse-starlette>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "httpx[http2]>=0.28.0",
]
serving = [
    "vllm>=0.6.0",
    "torch>=2.5.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "TCH"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_classes = "Test* *Test"
python_functions = "test_*"
filterwarnings = [
    "ignore::pytest.PytestCollectionWarning"
]

[tool.mypy]
python_version = "3.11"
strict = true
